requirements.txt
customtkinter>=5.2.0
anthropic>=0.18.0
reportlab>=4.0.0
Pillow>=10.0.0
